% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{cluster_label}
\alias{cluster_label}
\title{A rank based approach to modeling gene expression with filtered and normalized read count matrix}
\usage{
cluster_label(folder, mat, t = 3, name = "EXP", label_2)
}
\arguments{
\item{folder}{The normalised and filtered, read count matrix, with row names}

\item{mat}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{t}{Labels for the two sub-populations}

\item{name}{Labels for the two sub-populations}

\item{label_2}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}
}
\value{
label pValues A vector containing FDR adjusted p significance values
}
\description{
Takes in the complete filtered and normalized read count matrix, the location of the two sub-populations and the number of cores to be used
}
