% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{color_map}
\alias{color_map}
\title{finding colours for labels}
\usage{
color_map(labels)
}
\arguments{
\item{labels}{labels to which colours is required}
}
\value{
new_color would return list of colurs on the basis number labels'colour required
}
\description{
requirs labels to find colors for same
}
