% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{cosine_formula}
\alias{cosine_formula}
\title{cosine similarity formula}
\usage{
cosine_formula(A, B)
}
\arguments{
\item{A}{Second matrix}

\item{B}{First matrix}
}
\value{
cos_sim would return a matrix of pair-wise cosine similarity
}
\description{
requires two matrices to find pair-wise cosine similarity
}
