% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{matrix_filter}
\alias{matrix_filter}
\title{filtering matrices on the bass of common row and columns}
\usage{
matrix_filter(matrix_1, matrix_2, matrix_3, survival_info)
}
\arguments{
\item{matrix_1}{zfpkm single sell matrix}

\item{matrix_2}{tpm single cell matrix, same samples as in matrix_1}

\item{matrix_3}{tpm bulk TCGA matrix}

\item{survival_info}{survival info as two vector matrix, first vector as time and secondd as event}
}
\value{
filtered_matrix would return list of all matrices on common row as columns names
}
\description{
requires four matrices two for single cell and one for bulk, and  one for survival info to filter on same samples and features
}
