% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{num2bin}
\alias{num2bin}
\title{A rank based approach to modeling gene expression with filtered and normalized read count matrix}
\usage{
num2bin(mat, t = -3)
}
\arguments{
\item{mat}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{t}{Labels for the two sub-populations}
}
\value{
label pValues A vector containing FDR adjusted p significance values
}
\description{
Takes in the complete filtered and normalized read count matrix, the location of the two sub-populations and the number of cores to be used
}
