% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{ORsurv}
\alias{ORsurv}
\title{A rank based approach to modeling gene expression with filtered and normalized read count matrix}
\usage{
ORsurv(
  folder,
  matrix_1,
  matrix_2,
  matrix_3,
  t1 = -3,
  t2 = 3,
  t3 = 3,
  name1 = "EXP",
  name2 = "EXP_TCGA",
  name3 = "Survival_plot",
  survival_info,
  method,
  month_limit,
  p_limit = 2,
  selected_label
)
}
\arguments{
\item{folder}{The normalised and filtered}

\item{matrix_1}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{matrix_2}{Labels for the two sub-populations}

\item{matrix_3}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}

\item{t1}{Labels for the two sub-populations}

\item{t2}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}

\item{t3}{The number of cores to be used}

\item{name1}{The number of cores to be used}

\item{name2}{Labels for the two sub-populations}

\item{name3}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}

\item{survival_info}{The number of cores to be used}

\item{method}{The number of cores to be used}

\item{month_limit}{Labels for the two sub-populations}

\item{p_limit}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}

\item{selected_label}{The number of cores to be used}
}
\value{
surv_plot pValues A vector containing FDR adjusted p significance values
}
\description{
Takes in the complete filtered and normalized read count matrix, the location of the two sub-populations and the number of cores to be used
}
