% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{signature_matrix}
\alias{signature_matrix}
\title{A rank based approach to modeling gene expression with filtered and normalized read count matrix}
\usage{
signature_matrix(mat, label)
}
\arguments{
\item{mat}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{label}{for the two sub-populations}
}
\value{
sigmat pValues A vector containing FDR adjusted p significance values
}
\description{
Takes in the complete filtered and normalized read count matrix, the location of the two sub-populations and the number of cores to be used
}
