% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{survival_plot}
\alias{survival_plot}
\title{A rank based approach to modeling gene expression with filtered and normalized read count matrix}
\usage{
survival_plot(
  folder,
  mat,
  mat_2,
  s,
  t = 3,
  name2 = "EXP_TCGA",
  name3 = "Survival_plot",
  month_limit,
  p_limit,
  selected_label
)
}
\arguments{
\item{folder}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{mat}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{mat_2}{Labels for the two sub-populations}

\item{s}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}

\item{t}{Labels for the two sub-populations}

\item{name2}{Labels for the two sub-populations}

\item{name3}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}

\item{month_limit}{Labels for the two sub-populations}

\item{p_limit}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}

\item{selected_label}{The number of cores to be used}
}
\value{
surv_plot pValues A vector containing FDR adjusted p significance values
}
\description{
Takes in the complete filtered and normalized read count matrix, the location of the two sub-populations and the number of cores to be used
}
