% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{cluster_label}
\alias{cluster_label}
\title{find cluster's label on columns of matrix}
\usage{
cluster_label(folder = "pdf", mat, t = 3, name = "EXP")
}
\arguments{
\item{folder}{dir to save pdf files}

\item{mat}{matrix to clusters on columns}

\item{t}{number of desire clusters, dendrogram would cut at t depth}

\item{name}{name to save pdf file}
}
\value{
label would return cluster's label
}
\description{
It requires two matrix and number of clusters to cluster's label on columns of matrix
}
