% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{distance_matrix}
\alias{distance_matrix}
\title{finding pair-wise similarity of two matrices in three differnt ways}
\usage{
distance_matrix(mat_1, mat_2, method = "cosine")
}
\arguments{
\item{mat_1}{First matirx}

\item{mat_2}{Second matirx}

\item{method}{three options cosine, cor, and hamming default cosine}
}
\value{
mat_3 would return pair-wise similarity
}
\description{
requires two matrix and method to find pair-wise similarity
}
