% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{matrix_filter}
\alias{matrix_filter}
\title{A rank based approach to modeling gene expression with filtered and normalized read count matrix}
\usage{
matrix_filter(matrix_1, matrix_2, matrix_3, survival_info)
}
\arguments{
\item{matrix_1}{The normalised and filtered, read count matrix, with row names as genes name/ID and column names as sample id/name}

\item{matrix_2}{Labels for the two sub-populations}

\item{matrix_3}{Labels for the two sub-populations}

\item{survival_info}{optional, 10 as more accuracy with little slow (calculation on big number, Rmpfr library used), default is 0, means no Rmpfr library used for calculation}
}
\value{
filtered_matrix pValues A vector containing FDR adjusted p significance values
}
\description{
Takes in the complete filtered and normalized read count matrix, the location of the two sub-populations and the number of cores to be used
}
