% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{num2bin}
\alias{num2bin}
\title{numeric to binary conversion}
\usage{
num2bin(mat, t = -3)
}
\arguments{
\item{mat}{numeric matrix}

\item{t}{cutoff for 0 and 1}
}
\value{
mat would return binary matrix
}
\description{
Trequires matrix and cutoff as t for to convert numeric to binary matrix
}
