% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{ORsurv}
\alias{ORsurv}
\title{survival of patients using ors on TCGA data}
\usage{
ORsurv(
  folder,
  matrix_1,
  matrix_2,
  matrix_3,
  t1 = -3,
  t2 = 3,
  t3 = 3,
  name1 = "EXP",
  name2 = "EXP_TCGA",
  name3 = "Survival_plot",
  survival_info,
  method,
  month_limit,
  p_limit = 2,
  selected_label
)
}
\arguments{
\item{folder}{dir name to save pdf files}

\item{matrix_1}{zfpkm single sell matrix}

\item{matrix_2}{tpm single cell matrix, same samples as in matrix_1}

\item{matrix_3}{tpm bulk TCGA matrix}

\item{t1}{cutoff for 0 and 1}

\item{t2}{number of clusters on bianry matrix of single cell zfpkm}

\item{t3}{number of clusters on projected matrix of single cells and bulk cells}

\item{name1}{for binary matrix}

\item{name2}{for projected matrix plots}

\item{name3}{for survival plots}

\item{survival_info}{for survival info}

\item{method}{for pair-wise similarity method name}

\item{month_limit}{keeping survival info till specfic months, means number of months}

\item{p_limit}{therosld to selecte clusters where pair-wise p val is less}

\item{selected_label}{to pass already known cluster names out of all clusters}
}
\value{
surv_plot pValues A vector containing FDR adjusted p significance values
}
\description{
survival of patients using ors on TCGA data
}
