% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{signature_matrix}
\alias{signature_matrix}
\title{To find average of each label group}
\usage{
signature_matrix(mat, label)
}
\arguments{
\item{mat}{matrix for averaging}

\item{label}{a vector contains labels for each group}
}
\value{
sigmat return matrix of average vectors
}
\description{
It require number of column of mat equlal to label vector size
}
