% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORsurv.R
\name{survival_plot}
\alias{survival_plot}
\title{Survival plots}
\usage{
survival_plot(
  folder = "pdf",
  mat,
  mat_2,
  s,
  t = 3,
  name2 = "EXP_TCGA",
  name3 = "Survival_plot",
  month_limit,
  p_limit,
  selected_label
)
}
\arguments{
\item{folder}{dir name to save pdf files}

\item{mat}{prjected matrix getting after project of single cells and bulk TCGA data}

\item{mat_2}{Labels for the two sub-populations}

\item{s}{survival info}

\item{t}{number of cluster}

\item{name2}{specific for combine nameof single cell and bulk dataset names}

\item{name3}{name for survival plots}

\item{month_limit}{keeping survival info till specfic months, means number of months}

\item{p_limit}{therosld to selecte clusters where pair-wise p val is less}

\item{selected_label}{to pass already known cluster names out of all clusters}
}
\value{
surv_med would return median of each surv curve
}
\description{
requires projected matrix and survival info to plot survival and relevent plots
}
